refactor boilerplate-generator-tests
